PandoraCrypto module is dead
